# job-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Plateforme Emploi",
        "description": "Documentation Swagger des endpoints Admin pour la gestion des utilisateurs, offres et candidatures",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Lister toutes les offres d'emploi",
                "operationId": "61095dd3fdb7b25e7daa0084b396ca6e",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée des offres",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Job"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Créer une nouvelle offre",
                "operationId": "724b21b194916208376d71437be61a8b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JobRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Job"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/jobs/{id}": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Afficher une offre d'emploi",
                "operationId": "40e7eaa66d57253caa249e6f6fc06763",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du job",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Job"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Mettre à jour une offre",
                "operationId": "73ddd3048c72f5dbb787a4733069f5d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JobRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Job"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Supprimer une offre",
                "operationId": "7d6ee44631cf0dc1175069d51532c9b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supprimé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Job deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/employer/jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Lister les offres de l'employeur connecté",
                "operationId": "1b4c145d6c369d2bae47a7ff51ce3058",
                "responses": {
                    "200": {
                        "description": "Liste des offres",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Job"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Job": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "employer_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Développeur Laravel Senior"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description du poste..."
                    },
                    "company": {
                        "type": "string",
                        "example": "Tech Solutions"
                    },
                    "location": {
                        "type": "string",
                        "example": "Paris, France"
                    },
                    "salary": {
                        "type": "number",
                        "format": "float",
                        "example": 55000
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "full_time",
                            "part_time",
                            "contract",
                            "internship"
                        ],
                        "example": "full_time"
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "technology",
                            "healthcare",
                            "education",
                            "finance",
                            "other"
                        ],
                        "example": "technology"
                    },
                    "application_deadline": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-31"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "JobRequest": {
                "required": [
                    "title",
                    "description",
                    "company",
                    "location",
                    "type",
                    "category"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Développeur Laravel"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description détaillée du poste"
                    },
                    "company": {
                        "type": "string",
                        "example": "Tech Company"
                    },
                    "location": {
                        "type": "string",
                        "example": "Paris, France"
                    },
                    "salary": {
                        "type": "number",
                        "format": "float",
                        "example": 50000
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "full_time",
                            "part_time",
                            "contract",
                            "internship"
                        ],
                        "example": "full_time"
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "technology",
                            "healthcare",
                            "education",
                            "finance",
                            "other"
                        ],
                        "example": "technology"
                    },
                    "application_deadline": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-31"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}